fix modal behavior
